Refactor for cleaner dependencies
