+++ product_code/lagom/shared/src/main/scala/de/upb/cs/uc4/shared/SharedSerializerRegistry.scala
@@ -1,7 +1,8 @@
 package de.upb.cs.uc4.shared
 
 import com.lightbend.lagom.scaladsl.playjson.{JsonSerializer, JsonSerializerRegistry}
-import de.upb.cs.uc4.shared.messages.{Accepted, Confirmation, DetailedError, Rejected}
+import de.upb.cs.uc4.shared.api.{DetailedError, SimpleError}
+import de.upb.cs.uc4.shared.messages.{Accepted, Confirmation, Rejected}
 
 import scala.collection.immutable.Seq
 
@@ -10,9 +11,7 @@ trait SharedSerializerRegistry extends JsonSerializerRegistry {
   override def serializers: Seq[JsonSerializer[_]] = Seq(
     JsonSerializer[Confirmation],
     JsonSerializer[Accepted],
-    JsonSerializer[Rejected],
-    JsonSerializer[DetailedError],
-    JsonSerializer[Option[DetailedError]]
+    JsonSerializer[Rejected]
   ) ++ customSerializers
 
   /** All Service specific JsonSerializers */
