+++ product_code/lagom/shared_api/src/main/scala/de/upb/cs/uc4/shared/api/CustomException.scala
@@ -1,7 +1,6 @@
-package de.upb.cs.uc4.shared
+package de.upb.cs.uc4.shared.api
 
 import com.lightbend.lagom.scaladsl.api.transport.TransportErrorCode
-import de.upb.cs.uc4.shared.messages.DetailedError
 
 
 class CustomException(errorCode: TransportErrorCode, possibleErrorResponse: DetailedError, cause: Throwable) extends Exception(possibleErrorResponse.title, null, true, true){
