+++ product_code/lagom/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/UserServiceImpl.scala
@@ -12,8 +12,8 @@ import com.lightbend.lagom.scaladsl.persistence.{EventStreamElement, PersistentE
 import com.lightbend.lagom.scaladsl.server.ServerServiceCall
 import de.upb.cs.uc4.authentication.api.AuthenticationService
 import de.upb.cs.uc4.authentication.model.AuthenticationRole
-import de.upb.cs.uc4.shared.CustomException
 import de.upb.cs.uc4.shared.ServiceCallFactory._
+import de.upb.cs.uc4.shared.api.CustomException
 import de.upb.cs.uc4.shared.messages.{Accepted, Confirmation, Rejected, RejectedWithError}
 import de.upb.cs.uc4.user.api.UserService
 import de.upb.cs.uc4.user.impl.actor.{User, UserState}
